Move call recording widget types to type files and dedupe target lookup
Follow-up to #24810:
- Move CallRecordingWidgetTarget and CallRecordingWidgetUnavailableReason
  out of their hook files into dedicated files under the module's types
  folder, matching the exported-types convention
- Derive useCalendarEventTargetRecordId from useCallRecordingWidgetTarget
  so the widget target mapping lives in a single place
