+++ packages/twenty-front/src/modules/page-layout/widgets/call-recording/types/CallRecordingWidgetTarget.ts
@@ -1,0 +1,4 @@
+export type CallRecordingWidgetTarget = {
+  targetKind: 'calendarEvent' | 'callRecording';
+  recordId: string;
+};
