+++ packages/twenty-front/src/modules/page-layout/widgets/call-recording/types/CallRecordingWidgetUnavailableReason.ts
@@ -1,0 +1,3 @@
+export type CallRecordingWidgetUnavailableReason =
+  | 'workspaceWithoutCallRecording'
+  | 'recordWithoutCallRecording';
